requirements.txt
django-cookie-law
MySQL-python
PyMySQL
Pillow
